Clean up repo structure
